created basic dashboard that lists all plants assigned to a user.  also cleaned up the dashboard route in that there is only the home route now to reduce code duplication.  added new plant and edit plant routes
