+++ controllers/api/user-routes.js
@@ -60,19 +60,4 @@ router.post('/logout', (req, res) => {
   }
 });
 
-router.delete("/user/:id", (req, res) => {
-  User.destroy({
-    where: { id: req.params.id }
-  }).then(dbUserData => {
-    if (!dbUserData) {
-      res.status(404).json({ message: 'User does not exist.' });
-      return;
-    }
-  res.json(dbUserData);
-  }).catch(err => {
-    console.log(err);
-    res.status(500).json(err);
-  });
-});
-
 module.exports = router;
