+++ controllers/index.js
@@ -3,11 +3,9 @@ const router = require('express').Router();
 //require routes
 const apiRoutes = require('./api/');
 const homeRoutes = require('./home-routes.js');
-const dashboardRoutes = require('./dashboard-routes.js');
 
 //use routes
 router.use('/', homeRoutes);
-router.use('/dashboard', dashboardRoutes);
 router.use('/api', apiRoutes);
