+++ views/edit-plant.handlebars
@@ -1,0 +1,3 @@
+<div>
+    PLACEHOLDER : edit-plant ... needs work
+</div>
