--- views/layouts/dashboard.handlebars
@@ -1,3 +1,0 @@
-<div>
-    this is the dashboard ... it needs work
-</div>
