+++ views/new-plant.handlebars
@@ -1,0 +1,3 @@
+<div>
+    PLACEHOLDER : new-plant ... needs work
+</div>
